% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bitly_user_app_info.R
\name{bitly_app_details}
\alias{bitly_app_details}
\title{Retrieve details for the provided OAuth App client ID}
\usage{

  bitly_app_details(client_id = "be03aead58f23bc1aee6e1d7b7a1d99d62f0ede8",
  showRequestURL = F)
}
\arguments{
\item{client_id}{- The client ID of an OAuth app}
}
\description{
Retrieve details for the provided OAuth App client ID
}
