% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ApiKey.R
\name{bitly_auth}
\alias{bitly_auth}
\title{Assign API tokens using OAuth2.0}
\usage{
bitly_auth(key = "", secret = "")
}
\arguments{
\item{key}{- Client ID}

\item{secret}{- Client Secret}
}
\description{
You should register an application in order to get Client ID and Client Secret code. 
For Bit.ly, go to \url{https://bitly.com/a/oauth_apps} and in the field \code{Redirect URIs:} 
type for example "http://localhost:1410".
}
\examples{
\dontrun{
bitly_token <-
  bitly_auth(key = "be03aead58f23bc1aee6e1d7b7a1d99d62f0ede8",
             secret = "b7e4abaf8b26ec4daa92b1e64502736f5cd78899")
}

}
\seealso{
See \url{http://dev.bitly.com/rate_limiting.html}

See \url{http://dev.bitly.com/authentication.html}
}
