% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz.R
\name{bitly_bearerToken}
\alias{bitly_bearerToken}
\title{Set bit.ly bearer API key}
\usage{
bitly_bearerToken(token)
}
\description{
Set bearer token as env. variable.
}
\examples{
\dontrun{
bitly_bearerToken("access token")
}
}
\seealso{
\url{https://dev.bitly.com/docs/getting-started/authentication/}.
Inspired by \url{https://github.com/jhk0530/gemini.R}
}
\keyword{internal}
