% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bitly_bsd.R
\name{bitly_bsds}
\alias{bitly_bsds}
\title{Fetch all Branded Short Domains}
\usage{
bitly_bsds()
}
\description{
Fetch all Branded Short Domains
}
\seealso{
\url{https://dev.bitly.com/v4/#operation/getBSDs}
}
