% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bitly_camp.R
\name{bitly_create_campaigns}
\alias{bitly_create_campaigns}
\title{Create Campaign (Premium)}
\usage{
bitly_create_campaigns(
  group_guid = NULL,
  channel_guids = NULL,
  description = NULL,
  name = NULL,
  showRequestURL = T
)
}
\arguments{
\item{group_guid}{- a GUID for a Bitly group}

\item{channel_guids}{- a list of strings}

\item{description}{- description of campaign}

\item{name}{- its name}

\item{showRequestURL}{- an optional T/F value to whether show URL which has been
build and requested from server. For debug purposes, default FALSE.}
}
\description{
Create a new campaign
}
\section{Campaigns}{

Bitly Campaigns allows you to build and manage omnichannel campaigns.
}

\examples{
\dontrun{
cc <- bitly_create_campaigns(
  group_guid = "testing", showRequestURL = TRUE, channel_guids = list("1", "2", "3"),
  description = "description", name = "name"
)
}

}
\seealso{
\url{https://dev.bitly.com/api-reference/#createCampaign}
}
