% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bitly_camp.R
\name{bitly_create_channel}
\alias{bitly_create_channel}
\title{Create channel (Premium)}
\usage{
bitly_create_channel(
  access_token,
  group_guid = NULL,
  guid = NULL,
  name = NULL,
  modified = NULL,
  created = NULL,
  campaign_guid = NULL,
  bitlink_id = NULL,
  showRequestURL = T
)
}
\arguments{
\item{group_guid}{- a GUID for a Bitly group}

\item{guid}{- ID for a channel}

\item{name}{- its name}

\item{modified}{- string | ISO_TIMESTAMP}

\item{created}{- string | ISO TIMESTAMP}

\item{campaign_guid}{- string | A GUID for a Bitly campaign}

\item{bitlink_id}{- string}

\item{showRequestURL}{- an optional T/F value to whether show URL which has been
build and requested from server. For debug purposes, default FALSE.}
}
\description{
Create a new channel
}
\section{Campaigns}{

Bitly Campaigns allows you to build and manage omnichannel campaigns.
}

\examples{
\dontrun{
  gc <- bitly_create_channel(group_guid = "testing", ...)
}

}
\seealso{
\url{https://dev.bitly.com/api-reference/#createChannel}
}
