% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bitly_links.R
\name{bitly_expand_link}
\alias{bitly_expand_link}
\title{Expand a Bitlink}
\usage{
bitly_expand_link(access_token, bitlink_id = NULL, showRequestURL = FALSE)
}
\arguments{
\item{access_token}{- bearer token for authentication}

\item{bitlink_id}{- string}

\item{showRequestURL}{- an optional T/F value to whether show URL which has been
build and requested from server. For debug purposes, default FALSE.}
}
\value{
long_url - a full URL to which bitlink points to
}
\description{
See \url{https://dev.bitly.com/api-reference/#expandBitlink}
This endpoint returns public information for a Bitlink.
}
\examples{
\dontrun{
bitly_expand_link(bitlink_id = "bit.ly/DPetrov")
bitly_expand_link(bitlink_id = "on.natgeo.com/1bEVhwE")

## manyHashes <- list("bit.ly/DPetrov", "bit.ly/1QU8CFm", "bit.ly/1R1LPSE", "bit.ly/1LNqqva")
## for (u in 1:length(manyHashes)) {
##   print(bitly_expand_link(bitlink_id = manyHashes[[u]], showRequestURL = TRUE))
## }
}
}
