% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bitly_user_app_info.R
\name{bitly_rate_limits}
\alias{bitly_rate_limits}
\title{bitly_rate_limits}
\usage{
bitly_rate_limits(access_token, showRequestURL = F)
}
\arguments{
\item{access_token}{- bearer token for authentication}

\item{showRequestURL}{- an optional T/F value to whether show URL which has been
build and requested from server. For debug purposes, default FALSE.}
}
\value{
\code{data.frame} of end points and their rate limits by action
}
\description{
Provides bit.ly rate limits by endpoint. See \url{https://dev.bitly.com/api-reference/#getPlatformLimitss}
}
