% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bitly_links.R
\name{bitly_retrieve_bitlink}
\alias{bitly_retrieve_bitlink}
\title{Retrieve a Bitlink}
\usage{
bitly_retrieve_bitlink(bitlink = NULL, showRequestURL = FALSE)
}
\arguments{
\item{showRequestURL}{- an optional T/F value to whether show URL which has been 
build and requested from server. For debug purposes, default FALSE.}
}
\description{
This endpoint returns information for a Bitlink.
}
\examples{
\dontrun{
bitly_retrieve_bitlink(bitlink = "cnn.it/2HomWGB")
}

}
\seealso{
See \url{https://dev.bitly.com/v4/#operation/getBitlink}
}
