% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bitly_camp.R
\name{bitly_retrieve_campaigns}
\alias{bitly_retrieve_campaigns}
\title{Retrieve campaigns (Premium)}
\usage{
bitly_retrieve_campaigns(group_guid = NULL, showRequestURL = T)
}
\arguments{
\item{group_guid}{- a GUID for a Bitly group}

\item{showRequestURL}{- an optional T/F value to whether show URL which has been 
build and requested from server. For debug purposes, default FALSE.}
}
\description{
Retrieve campaigns (Premium)
}
\note{
It has not been tested, see readme file on GitHub \url{https://github.com/dmpe/urlshorteneR/}
Retrieve the campaigns for the current user
}
\examples{
\dontrun{
  gc <- bitly_retrieve_campaigns(group_guid = "testing")
}

}
\seealso{
\url{https://dev.bitly.com/v4/#operation/getCampaigns}
}
