% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bitly_camp.R
\name{bitly_retrieve_channels}
\alias{bitly_retrieve_channels}
\title{Retrieve channels (Premium)}
\usage{
bitly_retrieve_channels(
  access_token,
  group_guid = NULL,
  campaign_guid = NULL,
  showRequestURL = T
)
}
\arguments{
\item{group_guid}{- a GUID for a Bitly group}

\item{campaign_guid}{- string | A GUID for a Bitly campaign}

\item{showRequestURL}{- an optional T/F value to whether show URL which has been
build and requested from server. For debug purposes, default FALSE.}
}
\description{
Retrieve the channels available to a user
}
\section{Campaigns}{

Bit.ly Campaigns allows you to build and manage omnichannel campaigns.
}

\examples{
\dontrun{
gc <- bitly_retrieve_channels(group_guid = "testing", campaign_guid = "test")
}

}
\seealso{
\url{https://dev.bitly.com/api-reference/#getChannels}
}
