% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bitly_links.R
\name{bitly_retrieve_clicks_summary}
\alias{bitly_retrieve_clicks_summary}
\title{Get Clicks Summary for a Bitlink}
\usage{
bitly_retrieve_clicks_summary(bitlink = NULL, size = 50,
  unit_reference = NULL, unit = NULL, units = -1,
  showRequestURL = FALSE)
}
\arguments{
\item{size}{- The quantity of items to be be returned}

\item{unit}{- A unit of time}

\item{units}{- An integer representing the time units to query data for. pass -1 to return all units of time.}

\item{showRequestURL}{- show URL which has been build and requested from server. For debug
purposes.}
}
\description{
See \url{https://dev.bitly.com/v4/#operation/getClicksSummaryForBitlink}
This will return the click counts for a specified Bitlink. This rolls up all the data into a single field of clicks.
}
\examples{
\dontrun{
bitly_retrieve_clicks_summary(bitlink = "cnn.it/2HomWGB", unit = "day", units = -1, size = 100)
}
}
