% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bitly_custom_links.R
\name{bitly_retrieve_cust_bitlink}
\alias{bitly_retrieve_cust_bitlink}
\title{Retrieve Custom Bitlink (Premium)}
\usage{
bitly_retrieve_cust_bitlink(custom_bitlink = NULL,
  showRequestURL = FALSE)
}
\description{
Retrieve the details and history of a Custom Bitlink
}
\examples{
\dontrun{
bitly_retrieve_cust_bitlink(custom_bitlink = "es.pn/SuperBowl")
}

}
\seealso{
\url{https://dev.bitly.com/v4/#operation/getCustomBitlink}
}
