% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bitly_custom_links.R
\name{bitly_retrieve_cust_bitlink_metrics_destination}
\alias{bitly_retrieve_cust_bitlink_metrics_destination}
\title{Get Metrics for a Custom Bitlink by Destination (Premium)}
\usage{
bitly_retrieve_cust_bitlink_metrics_destination(
  custom_bitlink = NULL,
  showRequestURL = FALSE
)
}
\arguments{
\item{custom_bitlink}{- A Custom Bitlink made of the domain and keyword}

\item{showRequestURL}{- an optional T/F value to whether show URL which has been
build and requested from server. For debug purposes, default FALSE.}
}
\description{
Returns click metrics for the specified link by its historical destinations.
}
\examples{
\dontrun{
bitly_retrieve_cust_bitlink_metrics_destination(custom_bitlink = "es.pn/SuperBowl")
}

}
\seealso{
\url{https://dev.bitly.com/api-reference/#getCustomBitlinkMetricsByDestination}
}
