% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bitly_custom_links.R
\name{bitly_retrieve_destination_metrics}
\alias{bitly_retrieve_destination_metrics}
\title{Get Metrics for a Custom Bitlink by destination (Premium)}
\usage{
bitly_retrieve_destination_metrics(
  access_token,
  custom_bitlink = NULL,
  showRequestURL = FALSE
)
}
\arguments{
\item{access_token}{- bearer token for authentication}

\item{custom_bitlink}{- A Custom Bitlink made of the domain and keyword}

\item{showRequestURL}{- an optional T/F value to whether show URL which has been
build and requested from server. For debug purposes, default FALSE.}
}
\description{
Get Click Metrics for a Custom Bitlink by historical Bitlink destinations
}
\examples{
\dontrun{
bitly_retrieve_destination_metrics(custom_bitlink = "es.pn/SuperBowl")
}

}
\seealso{
\url{https://dev.bitly.com/api-reference/#getCustomBitlinkMetricsByDestination}
}
