% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bitly_groups.R
\name{bitly_retrieve_group}
\alias{bitly_retrieve_group}
\title{Retrieve a single group}
\usage{
bitly_retrieve_group(group_guid = NA, showRequestURL = T)
}
\arguments{
\item{group_guid}{- string | A GUID for a Bitly group}
}
\description{
Retrive details for a specific group that a user belongs to.
}
\section{Group}{

Groups are a subdivision within an organization. A user will belong to a group within an organization. 
Most actions on our API will be on behalf of a group. For example, when you 
shorten a link, it will be on behalf of a user and a group.
}

\examples{
\dontrun{
ui <- bitly_UserInfo(showRequestURL = TRUE) 
rg <- bitly_retrieve_group(group_guid = ui$default_group_guid)
}

}
\seealso{
\url{https://dev.bitly.com/v4/#operation/getGroup}
}
