% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bitly_groups.R
\name{bitly_retrieve_groups}
\alias{bitly_retrieve_groups}
\title{Retrieve a list of Groups}
\usage{
bitly_retrieve_groups(organization_guid = NA)
}
\arguments{
\item{organization_guid}{}
}
\description{
Retrive details for all groups that a user belongs to.
}
\section{Group}{

Groups are a subdivision within an organization. A user will belong to a group within an organization. 
Most actions on our API will be on behalf of a group. For example, when you 
shorten a link, it will be on behalf of a user and a group.
}

\seealso{
\url{https://dev.bitly.com/v4/#operation/getGroups}
}
