% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bitly_orgs.R
\name{bitly_retrieve_org_plan_limits}
\alias{bitly_retrieve_org_plan_limits}
\title{Get Plan Limits}
\usage{
bitly_retrieve_org_plan_limits(organization_id = NULL)
}
\arguments{
\item{organization_id}{- a required string | A GUID for a Bitly organization.
You may also simply pass "" (double quotes), but this should be avoided at all costs.}
}
\description{
Returns all plan limits and counts available for an organization.
}
\section{Organizations}{

Organizations are part of our hierarchy. This is the top level where a group and user will belong.
}

\examples{
\dontrun{
all_orgs <- bitly_retrieve_orgs()
plan <- bitly_retrieve_org_plan_limits(organization_id = all_orgs$guid)
}

}
\seealso{
\url{https://dev.bitly.com/api-reference/#getPlanLimits}
}
