% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bitly_orgs.R
\name{bitly_retrieve_organizations}
\alias{bitly_retrieve_organizations}
\title{Retrieve Organizations}
\usage{
bitly_retrieve_organizations()
}
\description{
Retrieve a list of organizations
}
\section{Organizations}{

Organizations are part of our hierarchy. This is the top level where a group and user will belong.
}

\examples{
\dontrun{
ros <- bitly_retrieve_organizations()
}

}
\seealso{
\url{https://dev.bitly.com/v4/#operation/getOrganizations}
}
