% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bitly_orgs.R
\name{bitly_retrieve_orgs}
\alias{bitly_retrieve_orgs}
\title{Retrieve all Organizations}
\usage{
bitly_retrieve_orgs()
}
\description{
Retrieve a list of organizations
}
\section{Organizations}{

Organizations are part of our hierarchy. This is the top level where a group and user will belong.
}

\examples{
\dontrun{
bitly_bearerToken("access token")
ros <- bitly_retrieve_orgs()
}
}
\seealso{
\url{https://dev.bitly.com/api-reference/#getOrganizations}
}
