% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bitly_camp.R
\name{bitly_update_campaign}
\alias{bitly_update_campaign}
\title{Update A Channel (Premium)}
\usage{
bitly_update_campaign(
  campaign_guid = NULL,
  group_guid = NULL,
  channel_guids = NULL,
  description = NULL,
  name = NULL,
  showRequestURL = T
)
}
\arguments{
\item{campaign_guid}{- string | A GUID for a Bitly campaign}

\item{group_guid}{- a GUID for a Bitly group}

\item{channel_guids}{- a list of strings}

\item{description}{- description of campaign}

\item{name}{- its name}

\item{showRequestURL}{- an optional T/F value to whether show URL which has been 
build and requested from server. For debug purposes, default FALSE.}
}
\description{
Update an existing Channel
}
\note{
It has not been tested, see readme file on GitHub \url{https://github.com/dmpe/urlshorteneR/}
}
\section{Campaigns}{

Bitly Campaigns allows you to build and manage omnichannel campaigns.
}

\examples{
\dontrun{
cc <- bitly_update_campaign(
  group_guid = "", channel_guids = list("1", "2", "3"),
  description = "description", name = "name"
)
}

}
\seealso{
\url{https://dev.bitly.com/v4/#operation/updateCampaign}
}
