% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bitly_custom_links.R
\name{bitly_update_cust_bitlink}
\alias{bitly_update_cust_bitlink}
\title{Update Custom Bitlink (Premium)}
\usage{
bitly_update_cust_bitlink(custom_bitlink = NULL, bitlink_id = NULL,
  showRequestURL = FALSE)
}
\arguments{
\item{bitlink_id}{- string}
}
\description{
Move a Keyword to a different Bitlink
}
\examples{
\dontrun{
bitly_update_cust_bitlink(custom_bitlink = "es.pn/SuperBowl", bitlink_id = "")
}

}
\seealso{
\url{https://dev.bitly.com/v4/#operation/updateCustomBitlink}
}
