% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bitly_information.R
\name{bitly_UserInfo}
\alias{bitly_UserInfo}
\title{Retrieve information for the current authenticated user}
\usage{
bitly_UserInfo(showRequestURL = FALSE)
}
\arguments{
\item{showRequestURL}{- show URL which has been build and requested from server. 
For debug purposes.}
}
\value{
login - the specified bitly login or the login of the authenticated user.

name - the user's full/display name

default_group_guid

created - Timestamp for the moment the user signed up.

is_created

modified

is_sso_user

is_2fa_enabled

email
}
\description{
See \url{https://dev.bitly.com/v4/#operation/getUser}
}
\note{
Both returned columns (!) are character type.
}
\examples{
\dontrun{ 
     bitly_token <- bitly_auth()
     ui <- bitly_UserInfo(showRequestURL = TRUE) 
}

}
