% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bitly_information.R
\name{is_user_premium_account_holder}
\alias{is_user_premium_account_holder}
\title{Check if authenticated user holds premium account}
\usage{
is_user_premium_account_holder()
}
\description{
Check if authenticated user holds premium account
}
