% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/LinksMetrics.R
\name{link.metrics.clicks}
\alias{link.metrics.clicks}
\title{Returns the number of clicks on a single Bitlink of the authenticated user.}
\usage{
link.metrics.clicks(link, limit = 1000, unit = c("minute", "hour", "day",
  "week", "month"), units = -1)
}
\arguments{
\item{link}{- a Bitlink.}

\item{limit}{- 1 to 1000 (default=1000).
##}

\item{unit}{- minute, hour, day, week or month, default: day; Note: when unit is minute the maximum value for units is 60.
value for each period of time.}

\item{units}{- an integer representing the time units to query data for. Pass -1 to return all units of time.}

\item{rollup}{- true (default) or false.  Return data for multiple units rolled up to a single result instead of a separate value for each period of time.}
}
\description{
See http://dev.bitly.com/link_metrics.html#v3_link_clicks
}
\examples{
link.metrics.clicks(link = "http://bit.ly/DPetrov", unit = "day", units = -1, limit = 100)
}

