% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/LinksMetrics.R
\name{link.metrics.encoders_by_count}
\alias{link.metrics.encoders_by_count}
\title{Returns users who have encoded this link (optionally only those in the requesting user's social graph), sorted by the number of clicks on each encoding user's link.}
\usage{
link.metrics.encoders_by_count(link, limit = 1000, my_network = "false",
  expand_user = "true", subaccounts = "false")
}
\arguments{
\item{link}{- a Bitlink.}

\item{limit}{- (optional) integer in the range 1:100 that specifies the number of records to return (default:10).}

\item{my_network}{(optional) true or false (default) - restrict to my network}

\item{expand_user}{(optional) true (default) or false - include display names of encoders}

\item{subaccounts}{(optional, only available to enterprise accounts) true or false (default) - restrict to this enterprise account and its subaccounts}
}
\value{
aggregate_link - the aggregate (global) Bitlink for the provided Bitlink.

entries - a mapping of link, user, and ts (when the Bitlink was created).
}
\description{
Returns users who have encoded this link (optionally only those in the requesting user's social graph), sorted by the number of clicks on each encoding user's link.
}
\note{
- The response will only contain users whose links have gotten at least one click, and will not contain any users whose links are private.
}
\examples{
rbitlyApi("0906523ec6a8c78b33f9310e84e7a5c81e500909")
link.metrics.encoders_by_count(link = "http://bit.ly/DPetrov", my_network = "false", expand_user = "true", limit = 100, subaccounts = "false")
}
\seealso{
See \url{http://dev.bitly.com/link_metrics.html#v3_link_encoders_by_count}
}

