% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/LinksMetrics.R
\name{link.metrics.referrers_by_domain}
\alias{link.metrics.referrers_by_domain}
\title{Returns metrics about the pages referring click traffic to a single Bitlink, grouped by referring domain.}
\usage{
link.metrics.referrers_by_domain(link, limit = 1000, unit = c("minute",
  "hour", "day", "week", "month"), units = -1)
}
\arguments{
\item{link}{- a Bitlink.}

\item{limit}{- 1 to 1000 (default=1000).}

\item{unit}{- minute, hour, day, week or month, default: day; Note: when unit is minute the maximum value for units is 60.
value for each period of time.}

\item{units}{- an integer representing the time units to query data for. Pass -1 to return all units of time.}
}
\value{
clicks - the number of clicks referred from this domain.

referrer - the URL referring clicks.
}
\description{
Returns metrics about the pages referring click traffic to a single Bitlink, grouped by referring domain.
}
\examples{
rbitlyApi("0906523ec6a8c78b33f9310e84e7a5c81e500909")
link.metrics.referrers_by_domain(link = "http://bit.ly/DPetrov", unit = "day", units = -1, limit = 100)
}
\seealso{
\url{http://dev.bitly.com/link_metrics.html#v3_link_referrers_by_domain}
}

