% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/bitly_linksMetrics.R
\name{link_Metrics_EncodersCount}
\alias{link_Metrics_EncodersCount}
\title{Returns the number of users who have shortened a link.}
\usage{
link_Metrics_EncodersCount(link, showRequestURL = FALSE)
}
\arguments{
\item{link}{- a Bitlink.}

\item{showRequestURL}{- show URL which has been build and requested from server. For debug purposes.}
}
\value{
aggregate_link - the aggregate (global) Bitlink for the provided Bitlink.

count - the number of bitly users who have shortened (encoded) this link.
}
\description{
Returns the number of users who have shortened (encoded) a single Bitlink.
}
\examples{
bitly_token <- bitly_auth(key = "be03aead58f23bc1aee6e1d7b7a1d99d62f0ede8", secret = "b7e4abaf8b26ec4daa92b1e64502736f5cd78899")
link_Metrics_EncodersCount(link = "http://bit.ly/DPetrov")
}
\seealso{
\url{http://dev.bitly.com/link_metrics.html#v3_link_encoders_count}
}

