% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/LinksMetrics.R
\name{link_Metrics_Owly}
\alias{link_Metrics_Owly}
\title{Given a owly URL, returns the number of clicks on it.}
\usage{
link_Metrics_Owly(shortUrl, from = "", to = "", showRequestURL = FALSE)
}
\arguments{
\item{shortUrl}{- a short URL}

\item{from}{- an optional string for specifing data range}

\item{to}{- an optional string for specifing data range}

\item{showRequestURL}{- show URL which has been build and requested from server. For debug purposes.}
}
\value{
Returns click data for the given time period on success.
}
\description{
Given an ow.ly URL, returns an array of dates and the number of clicks on that date.
The default behavior is to return all dates/clicks for that short URL. You can optionally specify
a date range to retrieve a subset of the data.
Date fields must be in the following format: YYYY-MM-DD HH:MM:SS
This is higly experimental ! Please report bugs if there are any.
}
\seealso{
See \url{http://ow.ly/api-docs#clickStats}
}

