% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/bitly_linksMetrics.R
\name{link_Metrics_ReferrersByDomain}
\alias{link_Metrics_ReferrersByDomain}
\title{Returns metrics about the pages referring click traffic to a single Bitlink, grouped by referring domain.}
\usage{
link_Metrics_ReferrersByDomain(link, limit = 1000, unit = c("minute",
  "hour", "day", "week", "month"), units = -1, showRequestURL = FALSE)
}
\arguments{
\item{link}{- a Bitlink.}

\item{limit}{- 1 to 1000 (default=1000).}

\item{unit}{- minute, hour, day, week or month, default: day; Note: when unit is minute the
maximum value for units is 60.}

\item{units}{- an integer representing the time units to query data for. Pass -1 to return all
units of time.}

\item{showRequestURL}{- show URL which has been build and requested from server. For debug purposes.}
}
\value{
clicks - the number of clicks referred from this domain.

referrer - the URL referring clicks.
}
\description{
Returns metrics about the pages referring click traffic to a single Bitlink, grouped by referring domain.
}
\examples{
bitly_token <- bitly_auth(key = "be03aead58f23bc1aee6e1d7b7a1d99d62f0ede8", secret = "b7e4abaf8b26ec4daa92b1e64502736f5cd78899")
link_Metrics_ReferrersByDomain(link = "http://bit.ly/DPetrov",unit="day",units=-1,limit = 100)
}
\seealso{
\url{http://dev.bitly.com/link_metrics.html#v3_link_referrers_by_domain}
}

