% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/Links.R
\name{links_clickStatsOwly}
\alias{links_clickStatsOwly}
\title{Given a owly URL, returns the number of clicks on it.}
\usage{
links_clickStatsOwly(shortUrl, from = "", to = "", showRequestURL = FALSE)
}
\arguments{
\item{shortUrl}{- a short URL}

\item{showRequestURL}{- show URL which has been build and requested from server. For debug purposes.}
}
\value{
Returns click data for the given time period on success.
}
\description{
Given an ow.ly URL, returns an array of dates and the number of clicks on that date.
The default behavior is to return all dates/clicks for that short URL. You can optionally specify
a date range to retrieve a subset of the data.
Date fields must be in the following format: YYYY-MM-DD HH:MM:SS
}
\seealso{
See \url{http://ow.ly/api-docs#clickStats}
}

