% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/Links.R
\name{links_ExpandOwly}
\alias{links_ExpandOwly}
\title{Given a owly URL, returns the target (long) URL.}
\usage{
links_ExpandOwly(shortUrl, showRequestURL = FALSE)
}
\arguments{
\item{shortUrl}{- a short URL to be expanded}

\item{showRequestURL}{- show URL which has been build and requested from server. For debug purposes.}
}
\value{
Returns the original url.
}
\description{
Given an ow.ly URL, returns the original full URL.
}
\seealso{
See \url{http://ow.ly/api-docs#expand}
}

