% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/Links.R
\name{links_InfoOwly}
\alias{links_InfoOwly}
\title{Given a owly URL, returns information about it.}
\usage{
links_InfoOwly(shortUrl, showRequestURL = FALSE)
}
\arguments{
\item{shortUrl}{- a short URL}

\item{showRequestURL}{- show URL which has been build and requested from server. For debug purposes.}
}
\value{
Returns full short url data on success.
}
\description{
Given an ow.ly URL, returns information about the page, including the original URL,
the HTML title, total clicks, and the "votes" value for the link (votes may be a positive or negative value).
}
\seealso{
See \url{http://ow.ly/api-docs#info}
}

