% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/Links.R
\name{links_Lookup}
\alias{links_Lookup}
\title{Query for a Bitlink based on a long URL.}
\usage{
links_Lookup(url, ..., showRequestURL = FALSE)
}
\arguments{
\item{url}{- one or more long URLs to lookup.}
}
\value{
url - an echo back of the url parameter.

aggregate_link - the corresponding bitly aggregate link (global hash).
}
\description{
Query for a Bitlink based on a long URL.
}
\examples{
rbitlyApi("0906523ec6a8c78b33f9310e84e7a5c81e500909")
links_Lookup(url = "http://www.seznam.cz/")
links_Lookup(url = "http://www.seznam.cz/", showRequestURL = TRUE)
}
\seealso{
See \url{http://dev.bitly.com/links.html#v3_link_lookup}
}

