% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/Links.R
\name{links_ShortenOwly}
\alias{links_ShortenOwly}
\title{Given a long URL, returns a Ow.ly link.}
\usage{
links_ShortenOwly(longUrl, showRequestURL = FALSE)
}
\arguments{
\item{longUrl}{- a long URL to be shortened (example: http://betaworks.com/).}

\item{showRequestURL}{- show URL which has been build and requested from server. For debug purposes.}
}
\value{
Returns short URL on success
}
\description{
Given a full URL, returns an ow.ly short URL. Currently the API only supports
shortening a single URL per API call. Your API Key controls whether the short url returned is
a static value (default), or whether it's always a unique value. See the Authentication \link{rbitlyApi} section
above for more details.
}
\seealso{
See \url{http://ow.ly/api-docs#shorten}
}

