% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bitly_orgs.R
\name{org_shorten_counts}
\alias{org_shorten_counts}
\title{Retrieve Organization Shorten Counts}
\usage{
org_shorten_counts(organization_guid = NULL)
}
\arguments{
\item{organization_guid}{- string | A GUID for a Bitly organization}
}
\description{
Retrieve all the shorten counts for a specific organization
}
\examples{
\dontrun{
osc <- org_shorten_counts(organization_guid = "")
}
}
\seealso{
\url{https://dev.bitly.com/v4/#operation/getOrganizationShortenCounts}
}
