% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/ApiKey.R
\name{rbitlyApi}
\alias{auth_bitly}
\alias{auth_owly}
\alias{rbitlyApi}
\title{Assign Bit.ly/Ow.ly API token}
\usage{
rbitlyApi(auth_token)

auth_bitly(auth_token)

auth_owly(x)
}
\arguments{
\item{auth_token/x}{- parameter to set Bit.ly Generic Access Token \code{\link{rbitlyApi}} and
or Ow.ly API key \url{http://ow.ly/api-docs}}
}
\description{
Assign Bit.ly/Ow.ly API token
}
\examples{
## rbitlyApi("0906523ec6a8c78b33f9310e84e7a5c81e500909") # now deprecated
options(Bit.ly = "0906523ec6a8c78b33f9310e84e7a5c81e500909", Ow.ly = "F1QH-Q64B-BSBI-JASJ") # use this for Bit.ly access
}
\seealso{
See \url{http://dev.bitly.com/rate_limiting.html}

See \url{http://dev.bitly.com/authentication.html}

See \url{http://ow.ly/api-docs}
}

