% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bitly_groups.R
\name{retrieve_group}
\alias{retrieve_group}
\title{Retrieve a Group}
\usage{
retrieve_group(group_guid = NA)
}
\arguments{
\item{group_guid}{- string | A GUID for a Bitly group}
}
\description{
Retrive details for a group
}
\examples{
\dontrun {
ui <- bitly_UserInfo(showRequestURL = TRUE) 
rg <- retrieve_group(group_guid = ui$default_group_guid)

}

}
\seealso{
\url{https://dev.bitly.com/v4/#operation/getGroup}
}
