% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/ApiKey.R
\name{returnApiKey}
\alias{returnApiKey}
\title{This method is for the case if user doesn't know what is his "Generic Access Token". When inserting username and password it will
return the key and assign it using \code{auth_token} method in the namespace}
\usage{
http://dev.bitly.com/rate_limiting.html
}
\arguments{
\item{username}{}

\item{password}{}
}
\description{
This method is for the case if user doesn't know what is his "Generic Access Token". When inserting username and password it will
return the key and assign it using \code{auth_token} method in the namespace
}
\examples{
{
returnApiKey("YourUsername", "YourPassword")
}
}

