% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/addins.R
\name{shortenerAddin}
\alias{shortenerAddin}
\title{Shorten an URL}
\usage{
shortenerAddin()
}
\description{
Call this function as an addin to prompt a url shortener.
  bit.ly accepts a custom domain, see \code{\link[urlshorteneR]{bitly_shorten_link}}.
}
