% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bitly_information.R
\name{update_user}
\alias{update_user}
\title{Changing your name or fetching basic user information}
\usage{
update_user(default_group_guid = "", name = "",
  showRequestURL = FALSE)
}
\arguments{
\item{name}{}
}
\description{
Changing your name or fetching basic user information
}
\note{
Applies only to the authenticated user.
}
