% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/UserInfo.R
\name{user.linkHistory}
\alias{user.linkHistory}
\title{Returns entries from a user's link history in reverse chronological order.}
\usage{
user.linkHistory(limit = 100)
}
\arguments{
\item{limit}{- optional integer in the range 1 to 100; default: 100, specifying the max number of results to return.}
}
\description{
See http://dev.bitly.com/user_info.html#v3_user_link_history
}

