% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/UserMetrics.R
\name{user.metrics.countries}
\alias{user.metrics.countries}
\title{Returns aggregate metrics about the countries referring click traffic to all of the authenticated user's Bitlinks.}
\usage{
user.metrics.countries(limit = 1000, unit = c("minute", "hour", "day",
  "week", "month"), units = -1)
}
\arguments{
\item{limit}{- 1 to 1000 (default=1000).}

\item{unit}{- minute, hour, day, week or month, default: day; Note: when unit is minute the maximum value for units is 60.
value for each period of time.}

\item{units}{- an integer representing the time units to query data for. Pass -1 to return all units of time.}
}
\value{
clicks - the number of clicks referred from this country.

country - the two-letter code of the referring country.
}
\description{
See http://dev.bitly.com/user_metrics.html#v3_user_countries
}
\note{
without the parameter unit this endpoint returns a legacy response format which assumes rollup=false, unit=day and units=7. When a unit is specified, rollup is always true.
}
\examples{
user.metrics.countries(unit = "day", units = -1, limit = 100)
}

