% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/UserMetrics.R
\name{user.metrics.popular_links}
\alias{user.metrics.popular_links}
\title{Returns the authenticated user's most-clicked Bitlinks (ordered by number of clicks) in a given time period.}
\usage{
user.metrics.popular_links(limit = 1000, unit = c("minute", "hour", "day",
  "week", "month"), units = -1)
}
\arguments{
\item{limit}{- 1 to 1000 (default=1000).}

\item{unit}{- minute, hour, day, week or month, default: day; Note: when unit is minute the maximum value for units is 60.
value for each period of time.}

\item{units}{- an integer representing the time units to query data for. Pass -1 to return all units of time.}
}
\value{
link - a Bitlink.

clicks - the number of clicks on that Bitlink in the specified timeframe.
}
\description{
Returns the authenticated user's most-clicked Bitlinks (ordered by number of clicks) in a given time period.
}
\note{
This replaces the realtime_links endpoint.
}
\examples{
rbitlyApi("0906523ec6a8c78b33f9310e84e7a5c81e500909")
user.metrics.popular_links(unit = "day", units = -1, limit = 100)
}
\seealso{
See \url{http://dev.bitly.com/user_metrics.html#v3_user_popular_links}
}

