% Generated by roxygen2 (4.1.1.9000): do not edit by hand
% Please edit documentation in R/UserMetrics.R
\name{user.metrics.shorten_counts}
\alias{user.metrics.shorten_counts}
\title{Returns the number of Bitlinks created in a given time period by the authenticated user.}
\usage{
user.metrics.shorten_counts(limit = 1000, unit = c("minute", "hour", "day",
  "week", "month"), rollup = c("false", "true"), units = -1)
}
\arguments{
\item{limit}{- 1 to 1000 (default=1000).}

\item{unit}{- minute, hour, day, week or month, default: day; Note: when unit is minute the
maximum value for units is 60.
value for each period of time.}

\item{rollup}{- true or false (default). Return data for multiple units rolled up to a single
result instead of a separate value for each period of time.}

\item{units}{- an integer representing the time units to query data for. Pass -1 to return all
units of time.}
}
\value{
tz_offset - the offset for the specified timezone, in hours.

unit - an echo of the specified unit value.

units - an echo of the specified units value.

user_shorten_counts - the number of shortens made by the specified user in the specified time.
}
\description{
Returns the number of Bitlinks created in a given time period by the authenticated user.
}
\examples{
rbitlyApi("0906523ec6a8c78b33f9310e84e7a5c81e500909")
user.metrics.shorten_counts(unit = "day", units = -1, limit = 100, rollup = "true")
user.metrics.shorten_counts(unit = "day", units = -1, limit = 100, rollup = "false")
user.metrics.shorten_counts(unit = "day", units = -1, limit = 100)
}
\seealso{
See \url{http://dev.bitly.com/user_metrics.html#v3_user_shorten_counts}
}

