% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/UserInfo.R
\name{user.tracking_domain_list}
\alias{user.tracking_domain_list}
\title{Returns a list of tracking domains a user has configured.}
\usage{
user.tracking_domain_list()
}
\value{
tracking_domains - a list of tracking domains configured for the authenticated user.
}
\description{
See http://dev.bitly.com/user_info.html#v3_user_tracking_domain_list
}

