% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/UserInfo.R
\name{user_TrackingDomains}
\alias{user_TrackingDomains}
\title{Returns a list of tracking domains a user has configured.}
\usage{
user_TrackingDomains(showRequestURL = FALSE)
}
\arguments{
\item{showRequestURL}{- show URL which has been build and requested from server. For debug purposes.}
}
\value{
tracking_domains - a list of tracking domains configured for the authenticated user.
}
\description{
Returns a list of tracking domains a user has configured.
}
\examples{
rbitlyApi("0906523ec6a8c78b33f9310e84e7a5c81e500909")
user_TrackingDomains()
}
\seealso{
See \url{http://dev.bitly.com/user_info.html#v3_user_tracking_domain_list}
}

