% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vgd_links.R
\name{vgd_LinksExpand}
\alias{vgd_LinksExpand}
\title{Expand a short URL to a longer one}
\usage{
vgd_LinksExpand(shorturl = "", showRequestURL = FALSE)
}
\arguments{
\item{shorturl}{- (optional) You can specify the shorturl parameter if you'd like to pick a 
shortened URL instead of having v.gd randomly generate one. These must be between 
5 and 30 characters long and can only contain alphanumeric characters and underscores. 
Shortened URLs are case sensitive. Bear in mind that a desired short URL might already 
be taken (this is very often the case with common words) so if you're using this option 
be prepared to respond to an error and get an alternative choice from your app's user.}

\item{showRequestURL}{- show URL which has been build and requested from server. 
For debug purposes.}
}
\description{
See \url{https://v.gd/apilookupreference.php}
}
\examples{
### isgd_LinksExpand(shorturl = "https://v.gd/4oIAXJ", showRequestURL = TRUE)

}
