% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bitly_bsd.R
\name{bitly_bsds}
\alias{bitly_bsds}
\title{Fetch all Branded Short Domains}
\usage{
bitly_bsds()
}
\description{
BSDs is an acronym for branded short domains. This is a custom 15
character or less domain for bitlinks. This allows you to customize the domain to your brand.
}
\seealso{
\url{https://dev.bitly.com/v4/#operation/getBSDs}
}
