% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bitly_camp.R
\name{bitly_create_channel}
\alias{bitly_create_channel}
\title{Create channel (Premium)}
\usage{
bitly_create_channel(group_guid = NULL, guid = NULL, name = NULL,
  modified = NULL, created = NULL, campaign_guid = NULL,
  bitlink_id = NULL, showRequestURL = T)
}
\arguments{
\item{group_guid}{- a GUID for a Bitly group}

\item{modified}{- string | ISO_TIMESTAMP}

\item{created}{- string | ISO TIMESTAMP}

\item{campaign_guid}{- string | A GUID for a Bitly campaign}

\item{showRequestURL}{- an optional T/F value to whether show URL which has been 
build and requested from server. For debug purposes, default FALSE.}
}
\description{
Create a new channel
}
\note{
It has not been tested, see readme file on GitHub \url{https://github.com/dmpe/urlshorteneR/}
}
\examples{
\dontrun{
  gc <- bitly_create_channel(group_guid = "testing", ...)
}

}
\seealso{
\url{https://dev.bitly.com/v4/#operation/createChannel}
}
