% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bitly_qr_codes.R
\name{bitly_qr_create_code}
\alias{bitly_qr_create_code}
\title{Create QR code with bit.ly link}
\usage{
bitly_qr_create_code(
  access_token,
  title = NULL,
  group_guid = NULL,
  bitly_link = NULL
)
}
\arguments{
\item{access_token}{- bearer token for authentication}

\item{title}{- a required string}

\item{group_guid}{- group id}

\item{bitly_link}{- a bit.ly link}
}
\description{
Create a new QR Code and return its metadata
}
\examples{
\dontrun{
  qr <- bitly_qr_create_code(group_guid = "Be2oejZbDDc", bitly_link = "bit.ly/abc1234")
}

}
\seealso{
\url{https://dev.bitly.com/api-reference/#createQRCodePublic}
}
