% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bitly_camp.R
\name{bitly_retrieve_campaign}
\alias{bitly_retrieve_campaign}
\title{Retrieve a Campaign}
\usage{
bitly_retrieve_campaign(campaign_guid = NULL, showRequestURL = T)
}
\arguments{
\item{campaign_guid}{- string | A GUID for a Bitly campaign}

\item{showRequestURL}{- an optional T/F value to whether show URL which has been 
build and requested from server. For debug purposes, default FALSE.}
}
\description{
Retrieve a Campaign
}
\note{
It has not been tested, see readme file on GitHub \url{https://github.com/dmpe/urlshorteneR/}
Retrive details for a campaign
}
\examples{
\dontrun{
  gc <- bitly_retrieve_campaign(campaign_guid = "testing")
}

}
\seealso{
\url{https://dev.bitly.com/v4/#operation/getCampaign}
}
