% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bitly_camp.R
\name{bitly_retrieve_campaigns}
\alias{bitly_retrieve_campaigns}
\title{Retrieve campaigns (Premium)}
\usage{
bitly_retrieve_campaigns(group_guid = NULL, showRequestURL = T)
}
\arguments{
\item{group_guid}{- a GUID for a Bitly group}

\item{showRequestURL}{- an optional T/F value to whether show URL which has been
build and requested from server. For debug purposes, default FALSE.}
}
\description{
Retrieve the campaigns for the current user
}
\section{Campaigns}{

Bit.ly Campaigns allows you to build and manage omnichannel campaigns.
}

\examples{
\dontrun{
gc <- bitly_retrieve_campaigns(group_guid = "testing")
}

}
\seealso{
\url{https://dev.bitly.com/api-reference/#getCampaigns}
}
