% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bitly_custom_links.R
\name{bitly_retrieve_cust_bitlink}
\alias{bitly_retrieve_cust_bitlink}
\title{Retrieve Custom Bitlink (Premium)}
\usage{
bitly_retrieve_cust_bitlink(custom_bitlink = NULL, showRequestURL = FALSE)
}
\arguments{
\item{custom_bitlink}{- A Custom Bitlink made of the domain and keyword}

\item{showRequestURL}{- an optional T/F value to whether show URL which has been
build and requested from server. For debug purposes, default FALSE.}
}
\description{
Retrieve the details and history of a Custom Bitlink
}
\examples{
\dontrun{
bitly_bearerToken("access token")
bitly_retrieve_cust_bitlink(custom_bitlink = "es.pn/SuperBowl")
}

}
\seealso{
\url{https://dev.bitly.com/api-reference/#getCustomBitlink}
}
