% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bitly_groups.R
\name{bitly_retrieve_groups}
\alias{bitly_retrieve_groups}
\title{Retrieve a list of all groups}
\usage{
bitly_retrieve_groups(access_token, organization_id = NULL, showRequestURL = F)
}
\arguments{
\item{organization_id}{- an optional string parameter | A GUID for a Bitly organization}

\item{showRequestURL}{- an optional T/F value to whether show URL which has been
build and requested from server. For debug purposes, default FALSE.}
}
\description{
Retrive details for all groups that a user belongs to.
}
\section{Group}{

Groups are a subdivision within an organization. A user will belong to a group within an organization.
Most actions on our API will be on behalf of a group. For example, when you
shorten a link, it will be on behalf of a user and a group.
}

\examples{
\dontrun{
rg <- bitly_retrieve_groups("") # will still work ok
}
}
\seealso{
\url{https://dev.bitly.com/api-reference#getGroups}
}
