% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bitly_orgs.R
\name{bitly_retrieve_organization}
\alias{bitly_retrieve_organization}
\title{Retrieve an Organization}
\usage{
bitly_retrieve_organization(organization_guid = NULL)
}
\arguments{
\item{organization_guid}{- string | A GUID for a Bitly organization}
}
\description{
Retrive details for an organization
}
\section{Organizations}{

Organizations are part of our hierarchy. This is the top level where a group and user will belong.
}

\examples{
\dontrun{
ro <- bitly_retrieve_organization(organization_guid = "")
}

}
\seealso{
\url{https://dev.bitly.com/v4/#operation/getOrganization}
}
