% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bitly_custom_links.R
\name{bitly_update_cust_bitlink}
\alias{bitly_update_cust_bitlink}
\title{Update Custom Bitlink (Premium)}
\usage{
bitly_update_cust_bitlink(
  custom_bitlink = NULL,
  bitlink_id = NULL,
  showRequestURL = FALSE
)
}
\arguments{
\item{custom_bitlink}{- A Custom Bitlink made of the domain and keyword}

\item{bitlink_id}{- string}

\item{showRequestURL}{- an optional T/F value to whether show URL which has been
build and requested from server. For debug purposes, default FALSE.}
}
\description{
Move a Keyword to a different Bitlink
}
\examples{
\dontrun{
bitly_update_cust_bitlink(custom_bitlink = "es.pn/SuperBowl", bitlink_id = "")
}

}
\seealso{
\url{https://dev.bitly.com/api-reference/#updateCustomBitlink}
}
