% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bitly_user_app_information.R
\name{bitly_update_user}
\alias{bitly_update_user}
\title{Update your name and/or default group ID

This will overwrite your (display) username and/or group ID you belong to.}
\usage{
bitly_update_user(default_group_guid = NULL, name = "",
  showRequestURL = FALSE)
}
\arguments{
\item{default_group_guid}{- group id}

\item{name}{- username}
}
\description{
Update your name and/or default group ID

This will overwrite your (display) username and/or group ID you belong to.
}
\note{
Applies only to the authenticated user:
Changing group ID is only permitted to premium users. Thus, if you are a "free" user and will try to change your
default group id to something else, you will get an error. In that case, only changing display name is permitted.
}
\examples{
\\donotrun {
# this applies only for "free" users
uu <- update_user(name= "Malc")

# if you are premium user, you can additionally adjust your group id
uug <- update_user(name= "Malc", default_group_guid = "TestGroupID")
}

}
\seealso{
\url{https://dev.bitly.com/v4/#operation/updateUser}
}
