% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/bitly_information.R
\name{bitly_UserTrackingDomains}
\alias{bitly_UserTrackingDomains}
\title{Returns a list of tracking domains a user has configured.}
\usage{
bitly_UserTrackingDomains(showRequestURL = FALSE)
}
\arguments{
\item{showRequestURL}{- show URL which has been build and requested from server.
For debug purposes.}
}
\value{
tracking_domains - a list of tracking domains configured for the authenticated user.
}
\description{
Returns a list of tracking domains a user has configured.
}
\examples{
\dontrun{
bitly_token <- bitly_auth(key = "be03aead58f23bc1aee6e1d7b7a1d99d62f0ede8", secret = "")
bitly_UserTrackingDomains()
}
}
\seealso{
See \url{http://dev.bitly.com/user_info.html#v3_user_tracking_domain_list}
}

