% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/ApiKey.R
\name{doRequest}
\alias{doRequest}
\title{Generalized function for executing GET Requests by appending user's Bit.ly API Key.}
\usage{
doRequest(url, authcode = rbitlyApi())
}
\arguments{
\item{url}{which is used for the request}

\item{authcode}{calls the rbitlyApi \code{\link{rbitlyApi}}}
}
\description{
Generalized function for executing GET Requests by appending user's Bit.ly API Key.
}

