% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/ApiKey.R
\name{google_auth}
\alias{auth_bitly}
\alias{auth_googl}
\alias{auth_owly}
\alias{bitly_auth}
\alias{google_auth}
\title{Assign Bit.ly/Ow.ly/Goo.gl API token or authenticate using OAUTH 2}
\usage{
google_auth(key = "", secret = "")

bitly_auth(key = "", secret = "")

auth_bitly(auth_token)

auth_owly(auth_token)

auth_googl(auth_token)
}
\arguments{
\item{key}{- Client ID}

\item{secret}{- Client Secret}

\item{auth_token}{- parameter to set Bit.ly Generic Access Token \code{\link{rbitlyApi}},
Ow.ly API key \url{http://ow.ly/api-docs} or Goo.gl API Key \url{https://console.developers.google.com/project}}
}
\description{
Assign Bit.ly/Ow.ly/Goo.gl API token or authenticate using OAUTH 2
}
\examples{
options(Bit.ly = "0906523ec6a8c78b33f9310e84e7a5c81e500909", Ow.ly = "F1QH-Q64B-BSBI-JASJ", Goo.gl = "")
## OR
google_auth(key = "806673580943-78jdskus76fu7r0m21erihqtltcka29i.apps.googleusercontent.com", secret = "qItL-PZnm8GFxUOYM0zPVr_t")
bitly_auth(key = "be03aead58f23bc1aee6e1d7b7a1d99d62f0ede8", secret = "b7e4abaf8b26ec4daa92b1e64502736f5cd78899")
}
\seealso{
See \url{http://dev.bitly.com/rate_limiting.html}

See \url{http://dev.bitly.com/authentication.html}

See \url{http://ow.ly/api-docs}

See \url{https://developers.google.com/url-shortener/v1/getting_started#APIKey}
}

