% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bitly_user_app_info.R
\name{is_bitly_user_premium_holder}
\alias{is_bitly_user_premium_holder}
\title{Check if authenticated user holds premium account}
\usage{
is_bitly_user_premium_holder()
}
\description{
Check if authenticated user holds premium account
}
\seealso{
[bitly_user_info()]
}
