% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/isgd_links.R
\name{isgd_LinksExpand}
\alias{isgd_LinksExpand}
\title{Expand a short URL to a longer one}
\usage{
isgd_LinksExpand(shorturl = "", showRequestURL = FALSE)
}
\description{
See \url{https://is.gd/apilookupreference.php}
}
\examples{
### isgd_LinksExpand(shorturl = "http://is.gd/4oIAXJ", showRequestURL = TRUE)

}

