% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/LinksMetrics.R
\name{link.metrics.encoders_count}
\alias{link.metrics.encoders_count}
\title{Returns the number of users who have shortened (encoded) a single Bitlink.}
\usage{
link.metrics.encoders_count(link)
}
\arguments{
\item{link}{- a Bitlink.}
}
\value{
aggregate_link - the aggregate (global) Bitlink for the provided Bitlink.

count - the number of bitly users who have shortened (encoded) this link.
}
\description{
Returns the number of users who have shortened (encoded) a single Bitlink.
}
\examples{
rbitlyApi("0906523ec6a8c78b33f9310e84e7a5c81e500909")
link.metrics.encoders_count(link = "http://bit.ly/DPetrov")
}
\seealso{
\url{http://dev.bitly.com/link_metrics.html#v3_link_encoders_count}
}

