% Generated by roxygen2 (4.1.1.9000): do not edit by hand
% Please edit documentation in R/ApiKey.R
\name{rbitlyApi}
\alias{rbitlyApi}
\title{Assign rbitly token}
\usage{
rbitlyApi(auth_token)
}
\arguments{
\item{auth_token}{- Passed parameter to set Bit.ly Generic Access Token \code{\link{rbitlyApi}}.}
}
\description{
Assign rbitly token
}
\examples{
rbitlyApi("0906523ec6a8c78b33f9310e84e7a5c81e500909")
}

