% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/ApiKey.R
\name{rbitlyApi}
\alias{rbitlyApi}
\title{Assign rbitly token automatically if I know it}
\usage{
rbitlyApi(auth_token)
}
\arguments{
\item{auth_token}{Optionally passed parameter to set Bit.ly Generic Access Token \code{\link{rbitlyApi}}.}
}
\value{
Returns invisibly the currently set rbitlyApi.
}
\description{
Assign rbitly token automatically if I know it
}
\examples{
rbitlyApi('foobar')
}

