% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bitly_camp.R
\name{retrieve_campaigns}
\alias{retrieve_campaigns}
\title{Retrieve campaigns (Premium)}
\usage{
retrieve_campaigns(group_guid = NULL, showRequestURL = T,
  verbose = T)
}
\arguments{
\item{group_guid}{- a GUID for a Bitly group}
}
\description{
Retrieve the campaigns for the current user
}
\examples{
\dontrun{
  gc <- retrieve_campaigns(group_guid = "testing")
}

}
\seealso{
\url{https://dev.bitly.com/v4/#operation/getCampaigns}
}
