% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bitly_camp.R
\name{retrieve_campaigns}
\alias{retrieve_campaigns}
\title{Retrieve campaigns (Premium)}
\usage{
retrieve_campaigns(group_guid = NULL, showRequestURL = T,
  verbose = T)
}
\arguments{
\item{group_guid}{- a GUID for a Bitly group}
}
\description{
Retrieve the campaigns for the current user
}
\examples{
\dontrun{
bitly_token <- bitly_auth(key = "206868fc5803d50c8d3aae9e7efb9e8a9c56067f", secret = "54987201d3599fd9a09cc80922394b0f250c7886")
gc <- retrieve_campaigns(group_guid = "testing")
}

}
\seealso{
\url{https://dev.bitly.com/v4/#operation/getCampaigns}
}
