% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/UserMetrics.R
\name{user.metrics.clicks}
\alias{user.metrics.clicks}
\title{Returns the aggregate number of clicks on all of the authenticated user's Bitlinks.
See http://dev.bitly.com/user_metrics.html#v3_user_clicks}
\usage{
user.metrics.clicks(link, limit = 1000, unit = c("minute", "hour", "day",
  "week", "month"), units = -1)
}
\description{
Returns the aggregate number of clicks on all of the authenticated user's Bitlinks.
See http://dev.bitly.com/user_metrics.html#v3_user_clicks
}
\note{
without the parameter unit this endpoint returns a legacy response format which assumes rollup=false, unit=day and units=7.
}
\examples{
user.metrics.clicks(link = "http://bit.ly/DPetrov", unit = "day", units = -1, limit = 100)
}

