% Generated by roxygen2 (4.1.1.9000): do not edit by hand
% Please edit documentation in R/UserMetrics.R
\name{user_Metrics_Referrers}
\alias{user_Metrics_Referrers}
\title{Returns aggregate metrics about the pages referring click traffic to all of the authenticated user's Bitlinks.}
\usage{
user_Metrics_Referrers(limit = 1000, unit = c("minute", "hour", "day",
  "week", "month"), rollup = c("false", "true"), units = -1)
}
\arguments{
\item{limit}{- 1 to 1000 (default=1000).}

\item{unit}{- minute, hour, day, week or month, default: day; Note: when unit is minute the
maximum value for units is 60.}

\item{rollup}{- true or false. Return data for multiple units rolled up to a single result
instead of a separate value for each period of time.}

\item{units}{- an integer representing the time units to query data for. Pass -1 to return
all units of time.}
}
\value{
clicks - the number of clicks referred from this URL.

referrer - the URL referring clicks.
}
\description{
Returns aggregate metrics about the pages referring click traffic to all of the authenticated user's Bitlinks.
}
\note{
When a unit is specified (always the case), rollup is always (!) true.
}
\examples{
rbitlyApi("0906523ec6a8c78b33f9310e84e7a5c81e500909")
user_Metrics_Referrers(unit = "day", units = -1, limit = 100, rollup = "true")
user_Metrics_Referrers(unit = "day", units = -1, limit = 100, rollup = "false")
}
\seealso{
See \url{http://dev.bitly.com/user_metrics.html#v3_user_referrers}
}

