% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/bitly_information.R
\name{user_TrackingDomains}
\alias{user_TrackingDomains}
\title{Returns a list of tracking domains a user has configured.}
\usage{
user_TrackingDomains(showRequestURL = FALSE)
}
\arguments{
\item{showRequestURL}{- show URL which has been build and requested from server.
For debug purposes.}
}
\value{
tracking_domains - a list of tracking domains configured for the authenticated user.
}
\description{
Returns a list of tracking domains a user has configured.
}
\examples{
bitly_token <- bitly_auth(key = "be03aead58f23bc1aee6e1d7b7a1d99d62f0ede8", secret = "b7e4abaf8b26ec4daa92b1e64502736f5cd78899")
user_TrackingDomains()
}
\seealso{
See \url{http://dev.bitly.com/user_info.html#v3_user_tracking_domain_list}
}

